#Pdf test utils!

+++ projects/ibs/pdf-problem-test/src/main/java/ru/ibs/pdf/problem/test/StartPdfTest.java
@@ -18,6 +18,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ArrayBlockingQueue;
@@ -38,8 +39,10 @@ public class StartPdfTest {
 
     public static void main(String[] args) throws DocumentException, IOException {
         System.out.println("Hello from pdf test!");
+        createFont();
         File jarPath = getPathToSelf();
         List<String> argsList = Arrays.asList(args);
+        deleteGoodFiles = deleteGoodFiles(argsList);
         File dir = getDir(argsList);
         if (!dir.exists()) {
             dir.mkdirs();
@@ -58,15 +61,18 @@ public class StartPdfTest {
             for (long i = 0L; i < maxFiles; i++) {
                 final long andIncrement = fileIndex.getAndIncrement();
                 final String threadName = "pdf_" + andIncrement;
-                final String[] executeParams;
+                final List<String> executeParams;
                 if (!isWindows()) {
-                    executeParams = new String[]{javaPath, "-jar", jarPath.getAbsolutePath(), "-d", dir.getAbsolutePath(), "-n", andIncrement + ""};
+                    executeParams = new ArrayList<>(Arrays.asList(javaPath, "-jar", jarPath.getAbsolutePath(), "-d", dir.getAbsolutePath(), "-n", andIncrement + ""));
                 } else {
-                    executeParams = new String[]{"cmd.exe", "/c", "start", "/wait", javaPath, "-jar", jarPath.getAbsolutePath(), "-d", dir.getAbsolutePath(), "-n", andIncrement + ""};
+                    executeParams = new ArrayList<>(Arrays.asList("cmd.exe", "/c", "start", "/wait", javaPath, "-jar", jarPath.getAbsolutePath(), "-d", dir.getAbsolutePath(), "-n", andIncrement + ""));
+                }
+                if (!deleteGoodFiles) {
+                    executeParams.add("-doNotDelete");
                 }
                 while (true) {
                     try {
-                        arrayBlockingQueue.put(executeParams);
+                        arrayBlockingQueue.put(executeParams.toArray(new String[1]));
                         break;
                     } catch (InterruptedException ex) {
                         ex.printStackTrace();
@@ -198,6 +204,10 @@ public class StartPdfTest {
         return args.contains("-master");
     }
 
+    private static boolean deleteGoodFiles(List<String> args) {
+        return !args.contains("-doNotDelete");
+    }
+
     private static long getFileNumber(List<String> args) {
         return Long.valueOf(args.get(args.indexOf("-n") + 1)).longValue();
     }
@@ -215,13 +225,15 @@ public class StartPdfTest {
     public static final String TIMES_FONT2 = "LiberationSans-Regular.ttf";
     private static Font baseFont;
 
-    static {
+    private static void createFont() throws RuntimeException {
         try {
-            baseFont = new Font(BaseFont.createFont(TIMES_FONT, ENCODING, BaseFont.EMBEDDED), 10);
+            BaseFont createFont = BaseFont.createFont(TIMES_FONT, ENCODING, BaseFont.EMBEDDED);
+            baseFont = new Font(createFont, 10);
             System.out.println(TIMES_FONT + " is used!");
         } catch (Exception e) {
             try {
-                baseFont = new Font(BaseFont.createFont(TIMES_FONT2, ENCODING, BaseFont.EMBEDDED), 10);
+                BaseFont createFont = BaseFont.createFont(TIMES_FONT2, ENCODING, BaseFont.EMBEDDED);
+                baseFont = new Font(createFont, 10);
                 System.out.println(TIMES_FONT2 + " is used!");
             } catch (Exception ex) {
                 try {
@@ -240,6 +252,8 @@ public class StartPdfTest {
         return paragraph;
     }
 
+    private static boolean deleteGoodFiles;
+
     private static void createTestPdf(File dir, long fileNumber) throws DocumentException, IOException {
         ByteArrayOutputStream stream = new ByteArrayOutputStream();
         Document document = new Document(PageSize.A4, 25, 25, 30, 25);
@@ -256,7 +270,7 @@ public class StartPdfTest {
         }
         Files.write(file.toPath(), byteArray, StandardOpenOption.CREATE_NEW);
         byte[] readAllBytes = Files.readAllBytes(file.toPath());
-        if (!isItErroneous(readAllBytes)) {
+        if (!isItErroneous(readAllBytes) && deleteGoodFiles) {
             file.delete();
         }
     }
